Add string as a source of iterator `for s <- 'abc'`
